+++ src/fplibclient.h
@@ -18,8 +18,8 @@ struct xyt_struct {
 
 #define FINGER_TEMPLATE_SIZE sizeof(xyt_struct)
 
-int get_fingerprint_template(const unsigned char* image_data, const unsigned int image_data_len, 
-	unsigned char* template_buf, const unsigned int template_buf_len);
+int get_fingerprint_template(const unsigned char* image_data, const size_t image_data_len, 
+	unsigned char* template_buf, const size_t template_buf_len);
 
 float cmp_fingerprint_template(void* tmp1, void* tmp2);
